third_party/WebKit/LayoutTests/vr/latest/exclusive_requestFrame_called.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/fake-vr-displays.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/vr_service.mojom.js"></script>
<script src="../resources/mock-vr-service.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();

vr_session_test( (t, session) => {
  function onFrame(vrFrame) {
    // TODO: Enable when VRPresentationFrame interface is added.
    /*t.step( () => {
      assert_true(vrFrame instanceof VRPresentationFrame);
    });*/
    t.done();
  }

  session.requestFrame(onFrame);
}, fakeDisplays["Pixel"], { exclusive: true },
"VRSession requestFrame properly calls the provided callback");

</script>
